Added bot perms checks

+++ modules/messages.py
@@ -18,29 +18,38 @@ async def purge(message: Message):
     if trans == 1:
         return
     try:
-        if message.reply_to_message:
-            member = await bot.get_chat_member(chat_id=message.chat.id,
-                                               user_id=message.from_user.id)
-            if member.status == 'creator' or member.status == 'administrator':
-                start = message.reply_to_message.message_id
-                end = message.message_id + 1
-                chat_id = message.chat.id
+        bot_obj = await bot.get_me()
+        bot_id = bot_obj.id
+        me = await bot.get_chat_member(chat_id=message.chat.id, user_id=bot_id)
+        if me.can_delete_messages:
+            if message.reply_to_message:
+                member = await bot.get_chat_member(chat_id=message.chat.id,
+                                                   user_id=message.from_user.id)
+                if member.status == 'creator' or member.status == 'administrator':
+                    start = message.reply_to_message.message_id
+                    end = message.message_id + 1
+                    chat_id = message.chat.id
 
-                msgs = []
-                for i in range(start, end):
-                    msgs.append(i)
-                    if len(msgs) == 100:
-                        await del_msgs(msgs, chat_id)
-                        msgs = []
+                    msgs = []
+                    for i in range(start, end):
+                        msgs.append(i)
+                        if len(msgs) == 100:
+                            await del_msgs(msgs, chat_id)
+                            msgs = []
 
-                await del_msgs(msgs, chat_id)
-                sent_msg = await bot.send_message(chat_id=chat_id, text=trans['messages']['purge'])
-                logger.info(f"{message.chat.full_name}: Purge completed")
-                sleep(5)
-                await bot.delete_message(chat_id=chat_id, message_id=sent_msg.message_id)
+                    await del_msgs(msgs, chat_id)
+                    sent_msg = await bot.send_message(chat_id=chat_id, text=trans['messages']['purge'])
+                    logger.info(f"{message.chat.full_name}: Purge completed")
+                    sleep(5)
+                    await bot.delete_message(chat_id=chat_id, message_id=sent_msg.message_id)
+            else:
+                await message.reply(trans['global']['errors']['no_reply'])
+                logger.warning(f'{message.chat.full_name}: User {message.from_user.full_name} tried to use command without reply')
         else:
-            await message.reply(trans['global']['errors']['no_reply'])
-            logger.warning(f'{message.chat.full_name}: User {message.from_user.full_name} tried to use command without reply')
+            perm = 'can_delete_messages'
+            await message.reply(trans['global']['errors']['no_needed_perm'].format(perm=perm),
+                                parse_mode='HTML')
+            logger.warning(f"{message.chat.full_name}: Bot doesn't have perm {perm}")
     except Exception as err:
         await message.reply(trans['global']['errors']['default'])
         logger.error(f"{message.chat.full_name}: User {message.from_user.full_name} {err}")
